EmojiLogs-App
Emoji Logs is designed as a emotional tracker. It records the emotions of the user, at request.
It records the location, text input from the user, date/time, and a photo of the location/object triggering the user's emotions, before creating a graph
based on the input.
The app is still in development, at this stage does the following:
- retrieves a location;
- presents a current location;
- takes input about emotions, from the user, based on emoticon image buttons;
- takes a photo of the location/object that has served as an emotional trigger;
- the fow of the app is controlled by timers.
The following step is to take input from the user, the page has already been constructed, however there is a lot of work beeing carried out by
the main thread and the app crashes.
A database will be designed to serve as a storage space for the app.
Intensive work is still required to achieve full functionality.