Miwok-translation app
A translation app created in Android Studio following a Android Studio tutorial step by step. The purpose of the app is to translate hard-coded, basic foreign words in english. It translates numbers, colours, basic words and small frases.
Based on Android Adapter the application is functional and it can be improved significantly.